+++ src/components/NavBar/NavBar.css
@@ -8,3 +8,8 @@
     align-items: center;
     gap: 8px;
 }
+.logo {
+    width: 100%;
+    height: 100%;
+    object-fit: cover
+}
