+++ src/pages/Architecture/ArchitecturesSection.css
