+++ src/pages/CanvasContent/index.js
@@ -1,0 +1,3 @@
+import CanvasContent from './CanvasContent';
+
+export default CanvasContent;
