+++ src/pages/Designer/index.js
@@ -1,0 +1,3 @@
+import Designer from "./Designer";
+
+export default Designer;
