+++ src/pages/Docs/index.js
@@ -1,0 +1,3 @@
+import DocHome from './DocHome';
+
+export default DocHome;
