--- src/pages/Flow.js
@@ -1,48 +1,0 @@
-import {
-    Box,
-    Step,
-    StepDescription,
-    StepIcon,
-    StepIndicator,
-    StepNumber,
-    StepSeparator,
-    StepStatus,
-    StepTitle,
-    Stepper,
-    useSteps,
-} from '@chakra-ui/react';
-import { render } from 'react-dom';
-
-const steps = [
-    { title: 'First', description: 'Contact Info' },
-    { title: 'Second', description: 'Date & Time' },
-    { title: 'Third', description: 'Select Rooms' },
-];
-
-function Example() {
-    const { activeStep } = useSteps({
-        index: 1,
-        count: steps.length,
-    });
-
-    return (
-        <Stepper index={activeStep}>
-            {steps.map((step, index) => (
-                <Step key={index}>
-                    <StepIndicator>
-                        <StepStatus complete={<StepIcon />} incomplete={<StepNumber />} active={<StepNumber />} />
-                    </StepIndicator>
-
-                    <Box flexShrink="0">
-                        <StepTitle>{step.title}</StepTitle>
-                        <StepDescription>{step.description}</StepDescription>
-                    </Box>
-
-                    <StepSeparator />
-                </Step>
-            ))}
-        </Stepper>
-    );
-}
-
-render(<Example />);
