+++ src/pages/Review/index.js
@@ -1,0 +1,3 @@
+import Review from './Review';
+
+export default Review;
