+++ src/pages/Wizard/Wizard.css
