+++ src/pages/Wizard/WizardBox.css
